+++ app.config.js
@@ -1,0 +1,48 @@
+const { EXPO_PUBLIC_SUPABASE_URL, EXPO_PUBLIC_SUPABASE_ANON_KEY } = process.env;
+
+module.exports = {
+  name: "GreenCompass",
+  slug: "GreenCompass",
+  version: "1.0.0",
+  orientation: "portrait",
+  icon: "./assets/images/icon.png",
+  scheme: "myapp",
+  userInterfaceStyle: "automatic",
+  newArchEnabled: true,
+  ios: {
+    supportsTablet: true,
+    bundleIdentifier: "com.anonymous.GreenCompass"
+  },
+  android: {
+    adaptiveIcon: {
+      foregroundImage: "./assets/images/adaptive-icon.png",
+      backgroundColor: "#ffffff"
+    },
+    package: "com.anonymous.GreenCompass"
+  },
+  web: {
+    bundler: "metro",
+    output: "static",
+    favicon: "./assets/images/favicon.png"
+  },
+  plugins: [
+    "expo-router",
+    [
+      "expo-splash-screen",
+      {
+        "image": "./assets/images/splash-icon.png",
+        "imageWidth": 200,
+        "resizeMode": "contain",
+        "backgroundColor": "#ffffff"
+      }
+    ]
+  ],
+  experiments: {
+    typedRoutes: true
+  },
+  // Add extra configuration for environment variables
+  extra: {
+    supabaseUrl: EXPO_PUBLIC_SUPABASE_URL,
+    supabaseAnonKey: EXPO_PUBLIC_SUPABASE_ANON_KEY,
+  }
+};